requirements.txt
attrs==19.3.0
tqdm==4.43.0
numpy==1.18.1
jinja2==2.11.1